Use relative paths, show flags in output
Need to use relative paths because the dependency tracking doesn't work on Windows if you use absolute paths. In MinGW, gcc generates the Makefile dep rules using Win32 paths (C:\Users\...) but make accepts Unix paths (/c/Users/...). This was causing make to not pick up changes made to header files and it took a FUCK OF A LOT of debugging to figure this out.