fix end of line behavior (fix #1)
The console will only advance to the next line when a character nees to be printed there. Most of the time, the cursor advances to the next cell immediately after a character is receieved and emitted. However, when a character is received while the cursor is placed in the last position of the line, the cursor will not advance to the next line. Instead, it will occupy the same cell as the received character until another character is received, at which point it will advance to the leftmost cell of the next line. This is done to avoid unnecessary blank lines and unwanted cursor movement when in 'raw' mode.