remove redundant multiply
multiply operates on byte operands, and kernel image occupies more clusters than can fit in a byte now. the operand being used to multiply lives in an area 1 byte wide and that cannot be change without messing up the FAT BPB, so we can't change the instruction to a word multiply. luckily the cluster size is one sector, so the multiply is actually unnecessary, thus we can remove it.